reestructure code
