+++ buyer-api/src/operations/addDataResponse.js
@@ -67,20 +67,3 @@ export const addDataResponse = async (dataOrder, dataResponse) => {
 
   return { id, status };
 };
-
-export const queueDataResponse = async (orderId, sellerAddress, sellerId) => {
-  const id = `${orderId}:${sellerAddress}`;
-
-  const existingDataResponse = await dataResponses.safeFetch(id);
-  if (existingDataResponse && existingDataResponse.status !== 'waiting') {
-    return { id, status: existingDataResponse.status };
-  }
-  if (!existingDataResponse) {
-    return { error: 'There\'s not a DataResponse whith the data provided' };
-  }
-
-  await dataResponses.update(id, { status: 'queued', sellerId });
-
-  await addProcessDataResponseJob({ orderId, dataResponseId: id });
-  return { id, status: 'queued' };
-};
