+++ buyer-api/src/utils/stores.js
@@ -105,3 +105,5 @@ export const buyerInfos = createLevelStore('buyer_infos');
 export const buyerInfoPerOrder = createLevelStore('buyer_info_per_order');
 export const notariesCache = createRedisStore('notaries.cache');
 export const notarizationRequests = createRedisStore('notarization_requests');
+/** @type {LevelStore<string, number>} */
+export const sellers = createLevelStore('sellers');
